+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased Changes
 
+- Added new tool `send_event` which allows sending an arbitrary event to Dynatrace
+- Added new scope `storage:events:write`
+
 ## 1.2.0
 
 **Note**: This release contains updates to vulnerable 3rd-party libraries `@modelcontextprotocol/sdk` (CVE-2026-0621) as well as `qs` (CVE-2025-15284). We advise to upgrade to this release as soon as possible.